pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pwnbrute"
description = "A small wrapper for probabilistic exploits based on pwntools"
license = { text = "MIT" }
authors = [
    { name = "Lnevx", email = "lnevx@protonmail.com" }
]
version = "1.1.0"
dependencies = ["pwntools<5,>=4.14"]

[tool.hatch.build.targets.sdist]
include = ["pwnbrute.py"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
preview = true
select = [
    "B",
    "A",
    "COM",
    "C4",
    "PIE",
    "RET",
    "SIM",
    "ARG",
    "PTH",
    "I",
    "N",
    "PERF",
    "E",
    "W",
    "D",
    "F",
    "PL",
    "TRY"
]
ignore = [
    "COM812",
    "D203",
    "D205",
    "D212",
    "TRY003"
]
per-file-ignores."examples/*.py" = [
    "D"
]
